requirements.txt
# Python 3.11 compatible requirements for Azure App Service
# Core dependencies with specific versions for deterministic builds
streamlit==1.28.2
openai==1.10.0
tiktoken==0.5.2

# LangChain ecosystem - let pip resolve compatible versions
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Vector database
chromadb==0.4.22
# pysqlite3-binary==0.5.2  # Only for Linux, commented for cross-platform compatibility

# Document processing
pypdf==3.17.4
python-docx==1.1.0
PyMuPDF==1.23.22
pillow==10.1.0

# Data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Azure storage
azure-storage-blob==12.19.0

# Async support
aiofiles==23.2.1

# Optional heavy dependencies (comment out if build fails)
# sentence-transformers==2.2.0
# faiss-cpu==1.7.4
# unstructured==0.10.30
# flashrank==0.2.5
# duckdb==0.10.0
# tabula-py==2.7.0
# pyprojroot==0.3.0